+++ packages/persona-kit/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add `launchedBy: "team-dispatcher"` to `AgentSpec` for dispatcher-launched team members.
+
 ## [3.0.42] - 2026-06-03
 
 ### Fixed